Fix macOS finding 6 – default input value missing
Problem: Simply pressing the enter key for e.g. "Enter your project: [cms]" does not take the default value.

Even on current macOS 14.5 Sonoma, /bin/bash is still version 3.2 without the readline
option -i (text to be placed into the buffer).

There was a separate implementation for Bash version < 4, but it did not work.
It is now reimplemented and always works without an input buffer and
without different implementations for the different bash versions.
The default value is displayed in square brackets and this default value is taken if no input was given.