Fix macOS finding 17 Cypress start + 13 again
The dix macOS finding 13 – sed Permission denied is working,
but it takes 2 minutes to compile new sed version.
And with joomla_cypress container and finding 17 Cypress is not starting
we have the problem again. Therefore changed to not use
option "-i" in all sed calls and deleted the upgrade-sed stuff.